fix hew lints by rust 1.66

+++ src/driver/btrfs.rs
@@ -8,7 +8,7 @@ pub(crate) fn gen_snapshot(cfg: &BtmCfg, idx: u64) -> Result<()> {
         return Err(eg!("Snapshot {} already exists!", idx));
     }
 
-    alt!(0 != (u64::MAX - idx) % cfg.itv as u64, return Ok(()));
+    alt!(0 != (u64::MAX - idx) % cfg.itv, return Ok(()));
     clean_outdated(cfg).c(d!())?;
     let cmd = format!(
         "
@@ -160,7 +160,7 @@ fn clean_outdated_fade(cfg: &BtmCfg) -> Result<()> {
         };
 
         pair.0.iter().for_each(|n| {
-            if 0 != (u64::MAX - n) % denominator as u64 {
+            if 0 != (u64::MAX - n) % denominator {
                 to_del.push(format!("{}@{}", &cfg.volume, n));
             }
         });
